requirements.txt
# Core environment
browsergym-core>=0.13.0
gymnasium>=0.29.0
playwright>=1.40.0

# ML / Training
torch>=2.1.0
transformers>=4.40.0
trl>=0.8.0
peft>=0.10.0
datasets>=2.18.0
accelerate>=0.28.0
bitsandbytes>=0.43.0

# Inference
vllm>=0.4.0

# API clients for MCTS data collection
anthropic>=0.25.0
openai>=1.12.0

# Utilities
numpy>=1.24.0
pyyaml>=6.0
tqdm>=4.66.0
pillow>=10.0.0
lxml>=5.0.0
